backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
python-dotenv
marshmallow
marshmallow-sqlalchemy
Flask-Cors